build succeeded

+++ Model/Entity.h
@@ -2,14 +2,13 @@
 #define ENTITY_H
 
 #include "structs.h"
-
-class Entity {
+class Entity{
 public:
     virtual ~Entity() {}
 
     // Pure virtual functions that must be implemented by any subclass
     virtual void attack() = 0;
-    virtual void takeDamage(int damage) = 0;
+    virtual void takeDamage(float damage) = 0;
     virtual coordinate getPosition() const = 0;
     virtual void setPosition(coordinate position) = 0;
     virtual void move(int deltaX, int deltaY) = 0;
